Add debug tag to DMA alloc prints
When the kernel module allocates some memory for the NPU to use, it
already prints a debug message with the size of the allocation. This
patch extends this to also print a short textual description of what
the allocation is for, for example "network-intermediate-data". This
is useful when debugging or analyzing.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>