Simplify Sram buffer deallocation
The logic used to determine when an SRAM buffer can be freed was
previously checking all outputs of each input to a node, to check if
that output had been prepared. It also had some logic to deal with
cases where multiple inputs to a node were produced by the same node.

This patch simplifies this logic by making use of the reference counting
feature of the SRAM allocator (this was not available at the time that
the code in question was originally written).

Note that a subtle change here is that SramAllocator now uses a vector
rather than set of 'users'. This allows the same node to have more than
one reference to a buffer and so removes the need for the uniqueness
check when looping over inputs. Although this makes lookups in the
data structure linear rather than constant, the size of the vector
will generally be very small (one or two). It will only be large if a
Node has a large number of consumers, which we do not see in any of our
supported networks.

The Compiler now calls PrepareAfterPassAssignment on all Nodes, rather
than just on the first Node in a Pass. This is necessary so that
PrepareAfterPassAssignment gets called on the last Node in a Pass,
which is the one that knows about the SRAM allocation.

This is a refactoring of existing code that should lead to no behavioral
change. Existing tests should therefore cover this.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>