Use DRAM buffer lifetime information for allocation
The DRAM buffer lifetime information is passed to the FirstFitAllocation
function which decides where to place buffers in an efficient manner.

Adds a debug dump of intermediate DRAM buffer lifetimes and addresses.

Adds a unit test for BufferManager to check that buffer alignment requirements are met.

Adds a component test for the Support Library to check that intermediate
buffer space is reused for a simple network.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>