Record lifetime of DRAM buffers
CompilerBufferInfo stores the start and end of each buffer's lifetime.
This is updated by calls to the new BufferManager::MarkBufferUsedAtTime
method.

The base Pass class calls this method as appropriate during generation,
to mark each input and output buffer of each Pass as being used, for
the range of command stream indexes generated for that Pass.

The lifetime information is not yet used, but a subsequent patch will
use this to allocate intermediate buffers more efficiently.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>