Change SramAllocator::Free to require success
Previously, SramAllocator::Free returned a bool to indicate whether the
operation was successful. This result was being ignored at all but one
call site, even though all the calls continued assuming that the
operation was successful.

There should never be a case where the operation can fail (unless there
is a programming error), so this patch changes the behaviour of Free,
so that it is undefined behaviour (i.e. it asserts) if the operation
failed. This means that the call sites do not need to add their own
individual checks.

A new TryFree method has been added with the old behaviour of Free
(i.e. it returns the success flag), in case a user does need to deal
with the failure case. This is currently used only in unit tests.

Also adds/improves the documentation on the SramAllocator methods.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>