Add FirstFitAllocation function
This allocates buffers based on their size and lifetime and is
implemented with a 'first-fit' scheme - each buffer is allocated at the
smallest memory address that gives a valid allocation (not overlapping
lifetime and memory with any other buffer). This is not an optimal
solution but it is quite fast and gives acceptable results for the use
case of intermediate DRAM buffer allocation.

This function is not used yet (except from unit tests), but this
function will be used in a future patch for intermediate DRAM
allocation.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>