Add Visitor functions for the Remaining Ple Operations
This patch adds support for LeakyRelu, Sigmoid, Tanh and
MeanXyOperations in the new NetworkToGraphOfPartsConverter.

New Visitor functions were implemented to handle LeakyRelu,
Sigmoid, Tanh and MeanXy Operations in a Network. The Visitors
generate the appropriate FusedPlePart types and append them to
the Graph Of Parts. Also, two new Unit Tests were added to test
the aforementioned functions.

Furthermore, a minor fix was applied to the AddIdentityWeights() in
FusedPlePart class to set the Input/Output quantization info.

Signed-off-by: Konstantinos Karavelas <konstantinos.karavelas@arm.com>