Enable cascading in Depthwise Convolution
Enables conversion of Depthwise Convolution operation to the Graph of
Parts and thus enabling cascading for the operation. This changes only
handles the basic version of Depthwise Convolution (stride: 1x1 and
channel multiplier: 1)

A test had also been created to check that the Parts created by the
NetworkToGraphOfParts converter are correct.

Signed-off-by: Eleanor Bonnici <eleanor.bonnici@arm.com>