Add FullyConnectedPart
FullyConnected needs to be added in cascading in order to support more
networks.

Add the FullyConnectedPart and plan generation with it. Only support
Lonely plans at the moment as fully connected layers are rare and
usually very large which means the likelihood they can be cascaded is
small and the impact is small.

Update McePart to remove logic to filter block configs for fully
connected as the McePart won't be used for FullyConnected operations.

Add a visitor function to the NetworkToGraphOfPartsConverter.

Use encoded weights instead of estimating the weight size in
estimation.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>