Fix Output Tensor calculation in strided Convolution
This patch fixes the Output Tensor calculation of the added
FusedPlePart, when there is a strided Convolution Operation in the
Network. The added FusedPlePart uses the Convolution's Input to
generate an intermediate interleave Tensor for its Output, which
is in turn passed as Input to the subsequent McePart.

Currently only a convolution stride of 2x2 is supported. In case of
a non-strided Convolution, the default Convolution Input is passed.

Signed-off-by: Konstantinos Karavelas <konstantinos.karavelas@arm.com>