Simplify EthosNPreCompiledObject in PreCompiledWorkload
As AddInput is called before AddOutput in CreateUncompiledNetwork
the order of inputs/outputs between Arm NN and ethos-n-backend can be
assumed to be the same. E.g. 0 -> 0, 1 -> 1, 2 -> 2 etc.

This allows the removal of the unordered_map's in the
EthosNPreCompiledObject, which makes it easier to cache a compiled
network and removes complexity.

However, if somehow the order isn't the same an exception is thrown.

Signed-off-by: Matthew Sloyan <matthew.sloyan@arm.com>