Add standalone PLE part
Standalone PLE part is added in cascading in order to support more
networks.

For addition and addition_rescale, only lonely plan is generated since
cascading for multiple input part is not supported.

For AVGPOOL_3X3_1_1_UDMA, a plan is generated only if the input and
output tensors can be fit into the SRAM.

New unit tests are created to validate this new feature.

Signed-off-by: Hui Chen <hui.chen@arm.com>