Add core address validation during device probing
If a core in the device tree does not have an address specified, a NULL
pointer will be returned when the resource holding the core's address is
requested. The returned resource pointer is currently not validated so
it gets passed on to the core probing, which crashes with a segmentation
fault because it tries to dereference the NULL pointer.

To handle this case gracefully, the device probing will now validate the
returned resource pointer before passing it to the core probing.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>