Remove resource NULL check in ethosn_map_iomem
The resource pointer NULL check before getting the resource's size in
the ethosn_map_iomem function doesn't add any protection because the
function will try to dereference the pointer regardless.

As this is an internal function, the caller should ensure that the
pointer is valid so the NULL check has been removed rather than adding
additional checks.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>