Adapt Pooling Visitor function for MAXPOOL Operations
This patch adds support for MAXPOOL_3X3_2_2_EVEN and
MAXPOOL_3X3_2_2_ODD in the new NetworkToGraphOfPartsConverter.

The Pooling Visitor function was modified to support the
aforementioned Operations, by checking the input tensor dimensions,
generating a FusedPlePart with MAXPOOL_3X3_2_2_EVEN/ODD in the
command_stream and appending the part to the GraphOfParts.
Also, a new Unit Test was added to test the modified Visitor
function.

Furthermore, a new lambda function for FusedPlePart creation was
implemented to simplify code. A minor fix was added for MeanXy to
work with different strides. The MeanXy Unit Test was modified to
check Parts generated by both MeanXy and Pooling Visitors.

Signed-off-by: Konstantinos Karavelas <konstantinos.karavelas@arm.com>