Account for upscaling in GetNumOperations
This function was previously using the input XY size to calculate the
number of multiplies and adds required, which therefore did not account
for any potential upscaling of the input performed by the MCE.

This patch changes the function to use the output XY size instead, which
will account for any upscaling.

This also fixes a similar bug for convolutions with "valid" padding,
where the IFM is larger than the OFM. The old code would have
overestimated the number of operations.

In all other cases, the results should be unchanged.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>