Move firmware and kernel module into Android vendor image
The firmware was in the wrong place from where the kernel module
looked for it, it is now copied with PRODUCT_COPY_FILES
and end up in the vendor partition when booting the

firmware_class.path=/vendor/lib/firmware/

Kernel boot arg can be used to specify this location.
Without it the firmware is looked for in /lib/firmware/
and this folder don't exist in Android.

BOARD_VENDOR_KERNEL_MODULES is also used to place the
kernel module in the correct place in the file system.
This fix the strange copy to the external kernel folder
and relying on other scripts to copy it back.

Signed-off-by: Zingo Andersen <zingo.andersen@arm.com>