Support Resize with cascading algorithm
The NetworkToGraphOfPartsConverter creates an McePart when it encounters
a Resize operation. This code is mostly a copy/paste of the legacy
NetworkToGraphConverter equivalent.

The McePart has new support for upscaling as part of the MCE operation
(technically it is the StripeGenerator that implements this support).

Rename the StripeGenerator::m_ShapeMultiplier field to
m_PleShapeMultiplier to distinguish it clearly from the new
m_MceShapeMultiplier field.

The McePart constructor already had a lot of parameters, so rather than
adding to this, this patch adds a struct which can be used to provide
the parameters. The old constructor has been left for
backwards-compatibility with other code.

MceOp has new fields for the upsample settings. These are not currently
required anywhere(as the estimation code needs only the tensor shapes),
however they will be required when we generate the command stream.

Adds unit tests for the NetworkToGraphOfPartsConverter and McePart plan
generation.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>