Support Relu with cascading algorithm
The NetworkToGraphOfPartsConverter modifies the previous McePart when it
encounters a Relu operation. If there is no previous McePart an
identity McePart is inserted.

Add CreateIdentityMcePart function and use it in Relu and Concat
Visitors.

BasePart now has HasActivationBounds and ModifyActivationBounds virtual methods.
McePart overloads them to modify the upper and lower bounds.

The McePart and MceOp now has support for upper and lower bounds

Adds unit tests for the NetworkToGraphOfPartsConverter.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>