Support for latest Arm NN
Arm NN have deprecated several methods on SubgraphView and added new
methods to be used instead. This patch changes some of our uses of
these methods to the new methods where it is easy to do so, and others
have been wrapped in ARMNN_NO_DEPRECATE_WARN_BEGIN/END to hide the
warnings.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>