Fix dangling references in unit tests
The Part types store references to some objects (e.g. EstimationOptions).
Some of the unit tests were passing in a reference to a local variable,
which meant that the Part object outlived the EstimationOptions it was
referring to.

This patch makes a simple fix to this, by allocating the objects at a
higher level in the call stack, so that they outlive the Part object.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>