Implement compilation caching for non-Android user
Added functionality that allows a user to pass SaveCachedNetwork and
CachedNetworkFilePath as model options to improve compilation time.

To enable saving SaveCachedNetwork should be set to true and a valid
file path to an empty file should be supplied for
CachedNetworkFilePath.

To enabling loading, SaveCachedNetwork should be set to false and valid
file path to the previously saved network should be supplied for
CachedNetworkFilePath. E.g. network.bin

Added EthosNCaching class to store cached networks temporarily.
This is stored as a shared pointer within EthosNCachingService which
acts as a Singleton. This allows it to be persistent across calls of
OptimizeSubgraphView as the EthosNBackend is created and destroyed
multiple times within Arm NN. Using this persistent object, the compiled
networks are added one by one and then is written to a file when saving.
When loading the file is read into this same object at the start
and is used to populate the compiled network blob without recompiling
which cuts down compilation time. This design was chosen for efficiency
as it allows writing and reading from the file to be done in one go,
rather than for each subgraph compilation.

Added EndToEnd unit tests.

Signed-off-by: Matthew Sloyan <matthew.sloyan@arm.com>