Improve mailbox error reporting and handling
Currently, only a subset of the mailbox messages have a status that can
be used to indicate if the message was successful or not. To improve
upon this, a new error response message has been added that can be used
to report an error for any message.

Errors reported with the new error response message, failing statuses in
old messages and failure to read from the mailbox are now seen as
unrecoverable and will cause the kernel to reset the NPU.

The firmware version has been given a minor version bump to indicate
support for the new error response message.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>