Fix issues with estimation for EstimateOnlyOps
Change the return value of EstimateOpGraph so that it can no longer
represent
"unestimated" Ops. The function now always handles all Ops, and if it
 doesn't then it is an error and it will throw an exception.

This is a partial revert of "Improve debugging for failed cascading
estimation" when this feature was added to the function, but it is no
longer necessary as the plan generation and estimation code are now
in-sync, so no OpGraphs are created which cannot be estimated.

This reduces confusion around EstimateOnlyOps, which do not get placed
into a Pass, but are handled correctly by the function.

Also fixes a bug where the failure reasons were not being returned from
the EstimateOpGraph function.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>