Access FullyConnected weights/bias in non-deprecated manner
The m_Weight and m_Bias members of FullyConnectedLayer are deprecated as
weights and bias are now separate ConstantLayers in the Graph. This
patch makes the backend use the new separate layers rather than the
deprecated members.

The AddBiases and AddWeights methods have been updated to be more
flexible by taking more parameters as input rather than assuming they
are members of the given Layer. This means they can be used for
both 'old-style' layers(where the weights/bias are members)
and 'new-style' layers (where the weights/bias are separate layers).
This should be valuable when Arm NN change other layers to the new
design.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>