Advertise ConstantTensorsAsInputs capability
This tells Arm NN that the Ethos-N backend supports the weights and bias
of Fully Connected being separate inputs to the layer.

A system test has been added to confirm that the backend correctly rejects
Fully Connected layers with non-constant weights, now that it is being
exposed to this possibility.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>