Fixed a merged buffer handling bug
Previously the cascading combiner wrongly connects a buffer that is
already merged with a buffer from previous part to the combined OpGraph.
This results in crashes when compiling some networks.

This is fixed by connecting the effective buffer to an operation in the
graph.

A unit test is added to avoid the regression.
Signed-off-by: Hui Chen <hui.chen@arm.com>