Fix requantization in Concat
When performing Concat operation the new tensor needs to be requantized
based on quantization of input tensors. This is done through calling
CreateIdentityMcePart(). The function took only inputQuantizationInfo as
an argument and could not therefore peform requantization.
It has been extended to also take outputQuantizationInfo as an argument
and then pass the relevant info to the MCE to perform the
requantization.

Signed-off-by: Eleanor Bonnici <eleanor.bonnici@arm.com>