Support Requantize with cascading algorithm
The NetworkGraphOfPartsConverter creates an McePart when it encounters a
Requantize operation and its input and output quantizations are
different. This is done by executing code for transpose convolution
with identity weights like in the current non-cascading graph compiler
flow. If the input and output quantizations are the same the operation
is ignored. ConnectNoop() function from the non-cascading flow has been
moved to the cascading flow and updated in order to connect the
operation's input and outputs when this is the case.

A unit test has been added to check the McePart is correctly created for
a single-layer Requantize network with different quantization of input
and output tensors.

Signed-off-by: Eleanor Bonnici <eleanor.bonnici@arm.com>