Fix order of passes in cascading estimation
The EstimateOpGraph function had several errors related to the order of
passes, when the passes were not created in topological order:
* The m_OpToPass map had incorrect pass indices
* The m_ParentIds were missing in some cases
* The m_OperationIds were wrong in some cases
* The passes were in the wrong order in some cases

This patch fixes these issues by deferring the creation of the final
list of passes until later in the function, once all Ops have been
processed. It also uses the order of Ops defined in the OpGraph rather
than the GetSortedOps function (which is now unused and has been removed).

Augment existing estimation tests to check the m_ParentIds and
m_OperationIds, and fix some incorrect expected results.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>