Show index of Ops in dot file dumps
The order of Ops within their OpGraph is important as it defines the
order of execution of the Ops. This information was not previously
visible in the dot file dumps, and this is useful to include for two
reasons:

1. In some cases of branching, the order may not be obvious from the
graph topology(e.g. which branch is executed first). 2. Even if the
order is obvious (e.g. a linear chain), a bug may mean that the order
actually doesn't match the topology.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>