Fix cascading GetParentIds for Ops not placed in a pass
GetParentIds for cascading was not behaving consistently with the
non-cascading equivalent for cases where Ops are not in any pass
(e.g. EstimateOnlyOps). This led to incorrect graphs being displayed in
SPA.

This patch changes the cascading version of the function to be behave
consistently with the non-cascading version and adds a unit test with a
case that was previously giving an inconsistent result.

The change from unordered_set to vector is to make the results
deterministic.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>