Fix bitmap extension not updating number of bits
During bitmap extension the number of bits were not updated and an
incorrect number of bits were copied from the original bitmap. This led
to leak of resources and iommu mapping conflicts.
Stream bitmap bits are now stored and used to retrieve the bitmap size
and to copy data to the extended bitmap.

Signed-off-by: Gabriele Vitiello <gabriele.vitiello@arm.com>