Fix intermittent failing test
The loop over plan.m_OutputMappings was not deterministic because
the order of the map depends on heap-allocated pointer values.
This led to Ops being added to the OpGraph in a non-deterministic order.

This patch changes the loop to iterate over the output slots, which
have a deterministic order. Unfortunately some additional logic was
needed to deal with duplicated output slots.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>