Fix incorrect printf formats in support queries
Using the signed int printf format (%d) for unsigned types, will cause
undefined behavior if the value does not fit into a signed int.

The instances of incorrect usage of the signed int format in the support
queries have now been replaced with the appropriate format. GCC's
compile-time format string check has been added to the SetReason
function so compiler warnings are raised if the wrong format is used.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>