Fixed an IFM SRAM tile corruption bug in cascading
The IFM tile size is clamped to the tensor size (rounded up to the
multiple of brick size). This can result in data corruption in SRAM. The
data in the input SRAM buffer for the plans belonging to the start of a
section or lonely part is from DRAM that could be FCAF compressed. The
HW always writes to the SRAM in full FCAF cell size and the tensor size
may not be multiple of FCAF cell size.

To fix this issue, the cascading compiler only clamp the IFM tile to the
tensor size for the plans that are in the middle or end of a section.
For the plans belonging to the beginging of a section, the IFM tile is
not clamped so that it would be FCAF aligned if the data is
FCAF decompressed.

A unit test is added to validate the IFM buffer size is correctly
reserved.

Signed-off-by: Hui Chen <hui.chen@arm.com>