Fix kernel module build for kernel 5.15 & 5.16
In kernel version 5.15, the dma_buf symbols have been moved to a
separate namespace so the kernel module needs to import the namespace to
be able to use the symbols.

In kernel version 5.16, the kernel build files have been changed to no
longer include header files from the compiler. This caused the kernel
module build to break because "stdbool.h" is no longer available. The
kernel provides a header with a typedef for bool so the inclusion has
been changed to use the kernel header instead.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>