Fixed double buffering of weights within a Section
Changed the double buffering logic in the Combiner to generate
Sections with double buffered weights, only when necessary. The
Section double buffers weights if the considered Parts allow for it,
otherwise it single buffers the weights.

This is achieved by creating a virtual function which determines
whether double buffering can be performed for each Part type. The
Combiner was modified to check and keep track of any double buffering
between Parts of the same Section. If double buffering occured in the
past, the next Part in the current Section will re-use the same
weight buffers as the previous set of Plans. Otherwise, it will check
whether it can single or double buffer and generate multiple Plans
for single or both single and double buffering weights.

Added 5x Unit Tests to check different sequences of Parts with and
without weights. A new Unit test framework was needed in order to
mock Parts with and without weights, forcing them to generate Plans
for selected cascadeTypes.

Signed-off-by: Konstantinos Karavelas <konstantinos.karavelas@arm.com>