Fix firmware overflowing the mailbox queue
A bug was introduced in commit "Update the queue write pointer after
pushing the header and payload" where the firmware could write data to
the response queue when there was no space left in the queue, resulting
in it overflowing and overwriting data that the kernel module was yet
to read. This could result in lost data and corruption.

This commit fixes the problem by changing ethosn_queue_write to use the
write pointer from the queue struct rather than taking it as a
parameter, which can be misused. In order to support writing multiple
parts of a message, it now accepts multiple buffers.

The write pointer is now updated only after the full message is written,
which simplifies the cache management on the writing side, and
simplifies the reading side as it does not need to deal with partial
messages.

ethosn_queue_read has been simplified by removing the output
read_pending parameter, as there are no more cases where the caller
needs to 'peek' data in the queue. ethosn_queue_write therefore updates
the read pointer directly.

Update firmware version number because the changes to the mailbox mean
that old versions of firmware/kernel module may not communicate
properly with the new versions.

Also remove validation of queue size which was being done on every use
of the queues, as this is already validated on initialisation.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>