Make cascading agents nested in XML command stream
Previously the XML command stream representation for cascading agents was:

<CASCADE>
   <NUM_AGENTS>6</NUM_AGENTS>
</CASCADE>
<AGENT>
   ...
</AGENT>
<AGENT>
   ...
</AGENT>
...

This was a bit confusing, because the agents "belong" to the CASCADE
command beforehand, but we weren't taking advantage of XML's hierarchical
format for this.

This patch changes the representation to be nested:

<CASCADE>
  <NUM_AGENTS>6</NUM_AGENTS>
  <AGENT>
     ...
  </AGENT>
  <AGENT>
     ...
  </AGENT>
  ...
</CASCADE>

Also split up the very long string literal in the unit tests as this
causes compilation issues on some compilers.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>