Implement the ProcessMceOp function
CascadingCompiler generates the cascading command stream by processing
each Op in the MergedOpGraph. This function process the Mce Ops, adds
the corresponding agents to the command stream, and adds the required
dependencies between the agents. As part of this change, some of the
already exisiting data structures has been renamed.

Signed-off-by: Mohamed Elzahhar <Mohamed.Elzahhar@arm.com>