Add unique ptr semantics to driver library Buffer
As the Buffer class' internal data can be different and handle different
kind of memory, both internally allocated and externally supplied we
want to avoid misuse and mistakes in the code. So the Buffer class is
updated to have the semantics of a unique_ptr, In other words
non-copyable, movable etc.

Signed-off-by: Zingo Andersen <zingo.andersen@arm.com>