Use utility Logger class in Driver library
Replace prints to std::cout and std::cerr with the logging framework
used in other parts of the driver stack. This provides more control
over enabling/disabling/filtering log messages etc.

In debug builds, messages are printed to stdout. In release builds
nothing is printed, to avoid unwanted log messages appearing in
partners' outputs. This behaviour is consistent with the Support
Library.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>