Use new Arm NN API and add import API
The import API is needed to support zero copy through Arm NN. We need to
use the new API in order to access the import API.

Created new functions for our tests which don't use the internal
Arm NN interfaces and just use the external interface

Caching tests caused following tests to fail as some stale state is
leftover. Fix this by saving the cache which resets the caching
singleton's state

Support the Import API by creating the EthosNImportTensorHandle and the
factory classes. Also added methods for needed functions in the
EthosNBackend and the EthosNPrecompiledWorkload

Input and output buffers in the precompiled workload are setup in
Execute(). It used to be done when the precompiled workload was created
but now it needs to done later as the tensor handles might not have
been created yet because they will be imported later.

Added a test to make sure Import API works by using a file descriptor
but it isn't a dma_buf file descriptor

Signed-off-by: Joshua Slater <joshua.slater@arm.com>