Add unit tests to check deallocation of Sram buffers
Two new unit tests were added in order to check Buffer allocation and
deallocation on a simple network. Plans are manually created using
Ops with Atomic and Cascade Lifetimes to test the SramAllocator
logic. Buffers consumed by Atomic Ops should be deallocated after
calling the corresponding CombinerDFS function.

A new member function was added in the SramAllocator to get the
number of allocations in Sram.

Signed-off-by: Konstantinos Karavelas <konstantinos.karavelas@arm.com>