Fix issue with not generating the correct plans for end of cascades
There was an issue where the end of a cascade was generating incorrect
plans for sections that used the full tensor. At the end of a cascade
we can split the output into multiple stripes to parallelise dmas and
compute. The plans generated used to split the memory buffer but not
the mce and ple compute stripes. This can't be executed.

Make sure the compute stripes are properly split and add a unit test.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>