Implement the ProcessDmaOp function
CascadingCompiler generates the cascading command stream by processing
each Op in the MergedOpGraph. This function process the DMA Ops, adds
the corresponding agents to the command stream, and adds the required
dependencies between the agents. As part of this change, the
MultiMapIterator is removed as it's no longer needed. Moreover, the
constexpr specifier is used to be more representative of the buffer
index in  different functions.

Signed-off-by: Mohamed Elzahhar <Mohamed.Elzahhar@arm.com>