Limit section size to agent window size
This change ensures that only sections of the right size are selected
by the Combiner. The Combiner discards combinations that exceed the
maximum number of agents supported by the firmware. A new field for
this maximum number has been added to firmware and hardware
capabilities. A new value has been assigned to agent window size in the
support library's internal implementation of firmware and hardware
capabilities.

Signed-off-by: Nabeel Ahmad <nabeel.ahmad@arm.com>