Fix uninitialized MceS in Cascading Compiler
The structs in the Cascading command stream does not have any
constructors so they need to be manually initialized or they will be
left containing indeterminate values.

The AddMceSchedulerToCommandStream function does not use all the fields
in the MceS struct so some are left with indeterminate values. To
resolve this, the MceS instance is now being zero-initialized before
being used.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>