Fix not passing device to tensor handle factories
The device model option that specifies which device that should be used
is never passed to the tensor handle factories so they assume that they
should use the default device. This causes issues on systems with
multiple devices because buffers created for one device cannot be used
on another.

To resolve this, the backend will now always check if a device has been
specified and if so pass it to the tensor handle factories.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>