Added Unit Test to check the Lifetime of Intermediate Dram Buffers
Added a new Unit Test that creates a network consisting of an
Intermediate Dram Buffer with multiple consumers. The Lifetime of the
Buffer dependes on the AgentId of the producer Op, as well as the Id
of its last consumer Op. That way, the Buffer will not be deallocated
prematurely.

New member functions were added to the CascadingCompiler to retrieve
the private members needed for Buffer Lifetime access.

Finally, a small fix was added in the IFM and OFM Streamers, to map
each Intermediate Dram Buffer to the respective BufferManager
BufferId.

Signed-off-by: Konstantinos Karavelas <konstantinos.karavelas@arm.com>