Fix uninitialized member in WeightStats
When a WeightStats struct is default constructed, its member variable is
not initialized and left with an indeterminate value. This happens
because the WeightStats inherits from the InputStats struct without
defining a new default constructor so only the constructor in InputStats
is called.

To resolve this, the WeightStats struct now has a default constructor to
initialize its member variable.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>