Change Glue structure and correct DMA transfer format for FC
Change the glue to be more flexible. Previously the glue didn't have a
concept of when it should be executed (e.g. at the end or the beginning
of plans), which impacted branches because you don't want to keep data
in SRAM for an entire branch.

There was a heuristic which added different parts of the glue to the
beginning or end of a plan. This wasn't flexible. There now needs to be
more complex glue for which include conversions (e.g. for fully
connected).

Overhaul the glue such that there is now explicit starting and ending
glue so the order of operations is explicit. And the connections
between plans and glues are explicit. Elems require both ending and
beginning glues for all part output and input slots.

Simplify how Combinations are constructed. There is now only 1 way to
construct a combination with a plan. The glue must be added explicitly
after.

Combinations which share part id's can't be added together anymore.

Fixed issue where StartSections isn't called with multiple inputs. This
should be possible e.g. starting a section with an Add. Remove the
ability to start sections for input and output plans as sections
require a previous SRAM buffer which input / output plans don't have.

Use the correct DMA transfer format for Fully Connected

Fully connected requires a non standard dma transfer as it uses NHWCB
transfer format on a reinterpreted NHWC tensor in order to DMA byte by
byte.

Be explicit with the transfer format for Dma and Concat Ops.

Include the NHWC dram buffer in the fully connected plan, this means
that we need to be able to generate glue between dram buffers.

Include the original input tensor shape for fully connected parts so it
can create the dram buffers needed.

Update tests to include the transfer formats.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>