Add support for Strided Convolution in Cascading Compiler
Cascading compiler needs to fill the MceScheduler agent data to enable
the support for strided convolution. This patch sets the proper filter
shapes, padding, and IFM delta information for that purpose. As part of
this patch, the NetworkToGraphOfPartsConverter and McePart has been
updated to add and store the uninterleaved input tensor shape.

Signed-off-by: Mohamed Elzahhar <Mohamed.Elzahhar@arm.com>