Add support for Upsampling in Cascading Firmware
- Upsampling support in cascading firmware requires additional
information about the upsampling type and the edge upsampling mode. This
information is provided to the cascading firmware by adding two new
fields to MceS in the cascading command stream.

- Until now, the assumption that IFM stripe shape is same under all
scenarios was being made for MceS. This is not true as an edge IFM
stripe's dimension might differ w.r.t. other (default) IFM stripes. To
fix this, an additional field that describes MceS's IFM edge stripe
shape has been added to the cascading command stream.

Signed-off-by: Shubham <shubham@arm.com>