Fix glue for branches in the combiner
A branch cannot share glue with other branches if (1) it is an output
(2) or it is input format in DRAM is NHWC.

This will prevent the combiner from sharing intermediate and output
buffers and support fully connected whose input is in NHWC format.

Several combiner glue unit tests are modified to reflect the change. A
new unit test is added to validate the case where branches have mixed
formats.

Signed-off-by: Hui Chen <hui.chen@arm.com>